+++ test/e2e/tests/network/network-manager.spec.ts
@@ -11,7 +11,7 @@ import { withFixtures } from '../../helpers';
 import { login } from '../../page-objects/flows/login.flow';
 import SelectNetworkModal, {
   NetworkId,
-} from '../../page-objects/pages/dialog/select-network-modal';
+} from '../../page-objects/pages/networks/select-network-modal';
 import NetworkFilter from '../../page-objects/pages/networks/network-filter';
 import TokensTab from '../../page-objects/pages/home/tokens-tab';
 import TestDapp from '../../page-objects/pages/test-dapp';
