+++ test/e2e/tests/request-queuing/dapp1-switch-dapp2-send.spec.ts
@@ -9,7 +9,7 @@ import {
 import ActivityTab from '../../page-objects/pages/home/activity-tab';
 import ConnectAccountConfirmation from '../../page-objects/pages/confirmations/connect-account-confirmation';
 import HomePage from '../../page-objects/pages/home/homepage';
-import SelectNetworkModal from '../../page-objects/pages/dialog/select-network-modal';
+import SelectNetworkModal from '../../page-objects/pages/networks/select-network-modal';
 import NetworkFilter from '../../page-objects/pages/networks/network-filter';
 import NetworkPermissionSelectModal from '../../page-objects/pages/dialog/network-permission-select-modal';
 import ReviewPermissionsConfirmation from '../../page-objects/pages/confirmations/review-permissions-confirmation';
