+++ test/e2e/tests/tokens/import-tokens.spec.ts
@@ -15,7 +15,7 @@ import {
 import { mockTokensV2SupportedNetworks } from '../btc/mocks/tokens-api';
 import SelectNetworkModal, {
   NetworkId,
-} from '../../page-objects/pages/dialog/select-network-modal';
+} from '../../page-objects/pages/networks/select-network-modal';
 import NetworkFilter from '../../page-objects/pages/networks/network-filter';
 import { getMockAssetsPrice } from './utils/mocks';
