+++ test/e2e/tests/tokens/nft/filter-nfts.spec.ts
@@ -7,7 +7,7 @@ import Homepage from '../../../page-objects/pages/home/homepage';
 import NftsTab from '../../../page-objects/pages/home/nfts-tab';
 import SelectNetworkModal, {
   NetworkId,
-} from '../../../page-objects/pages/dialog/select-network-modal';
+} from '../../../page-objects/pages/networks/select-network-modal';
 import NetworkFilter from '../../../page-objects/pages/networks/network-filter';
 import { login } from '../../../page-objects/flows/login.flow';
 import { NETWORK_CLIENT_ID } from '../../../constants';
