+++ test/e2e/tests/tron/network.spec.ts
@@ -6,7 +6,7 @@ import { Driver } from '../../webdriver/driver';
 import { login } from '../../page-objects/flows/login.flow';
 import HomePage from '../../page-objects/pages/home/homepage';
 import TokensTab from '../../page-objects/pages/home/tokens-tab';
-import SelectNetworkModal from '../../page-objects/pages/dialog/select-network-modal';
+import SelectNetworkModal from '../../page-objects/pages/networks/select-network-modal';
 import NetworkFilter from '../../page-objects/pages/networks/network-filter';
 import HeaderNavbar from '../../page-objects/pages/header-navbar';
 import NetworksPage from '../../page-objects/pages/networks/networks-page';
